requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
rembg>=2.0.67
onnxruntime>=1.16.0
Pillow>=10.2.0
python-dotenv>=1.0.0
pydantic>=2.5.0

# Testing dependencies
pytest>=7.4.3
pytest-asyncio>=0.21.1
httpx>=0.25.2
pytest-mock>=3.12.0
pytest-cov>=4.1.0
aiohttp>=3.8.0  # For load testing
